Completed first testing of new sync
