+++ beacon_node/network/src/sync/range_sync/chain.rs
@@ -1,3 +1,4 @@
+use crate::sync::message_processor::{PeerSyncInfo, FUTURE_SLOT_TOLERANCE};
 use crate::sync::network_context::SyncNetworkContext;
 use beacon_chain::{BeaconChain, BeaconChainTypes, BlockProcessingOutcome};
 use eth2_libp2p::rpc::methods::*;
@@ -5,6 +6,7 @@ use eth2_libp2p::rpc::RequestId;
 use eth2_libp2p::PeerId;
 use fnv::FnvHashMap;
 use slog::{crit, debug, info, trace, warn, Logger};
+use std::cmp::Ordering;
 use std::collections::HashSet;
 use std::ops::Sub;
 use std::sync::Weak;
@@ -17,12 +19,14 @@ use types::{BeaconBlock, EthSpec, Hash256, Slot};
 /// to do so.
 //TODO: Make this dynamic based on peer's bandwidth
 const BLOCKS_PER_REQUEST: u64 = 50;
+
+#[derive(PartialEq)]
 pub struct Batch<T: EthSpec> {
     /// The ID of the batch, batches are ID's sequentially.
     id: u64,
-    /// The requested start slot of the batch.
+    /// The requested start slot of the batch, inclusive.
     start_slot: Slot,
-    /// The requested end slot of batch.
+    /// The requested end slot of batch, exclusive.
     end_slot: Slot,
     /// The hash of the chain root to requested from the peer.
     head_root: Hash256,
@@ -35,6 +39,20 @@ pub struct Batch<T: EthSpec> {
     /// The blocks that have been downloaded.
     downloaded_blocks: Vec<BeaconBlock<T>>,
 }
+
+impl<T: EthSpec> Ord for Batch<T> {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.id.cmp(&other.id)
+    }
+}
+
+impl<T: EthSpec> PartialOrd for Batch<T> {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl<T: EthSpec> Eq for Batch<T> {}
 
 impl<T: EthSpec> Batch<T> {
     fn new(id: u64, start_slot: Slot, end_slot: Slot, head_root: Hash256, peer_id: PeerId) -> Self {
@@ -124,8 +142,8 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
             pending_batches: FnvHashMap::default(),
             completed_batches: Vec::new(),
             peer_pool,
-            to_be_downloaded_id: 0,
-            to_be_processed_id: 0,
+            to_be_downloaded_id: 1,
+            to_be_processed_id: 1,
             last_processed_id: 0,
             state: ChainSyncingState::Stopped,
         }
@@ -175,6 +193,7 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
         // An entire batch of blocks has been received. This functions checks to see if it can be processed,
         // remove any batches waiting to be verified and if this chain is syncing, request new
         // blocks for the peer.
+        warn!(log, "Completed batch received"; "id"=>batch.id, "blocks"=>batch.downloaded_blocks.len(), "awaiting_batches" => self.completed_batches.len());
 
         // The peer that completed this batch, may be re-requested if this batch doesn't complete
         // the chain and there is no error in processing
@@ -201,9 +220,8 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
 
         let insert_index = self
             .completed_batches
-            .iter()
-            .position(|iter_batch| iter_batch.id > batch.id)
-            .unwrap_or_else(|| 0);
+            .binary_search(&batch)
+            .unwrap_or_else(|index| index);
         self.completed_batches.insert(insert_index, batch);
 
         // We have a list of completed batches. It is not sufficient to process batch successfully
@@ -213,7 +231,14 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
         // already be processed but not verified and therefore have Id's less than
         // `self.to_be_processed_id`.
 
+        //TODO: Run the processing of blocks in a separate thread. Build a queue of completed
+        //blocks here, manage the queue and process them in another thread as they become
+        //available.
+
         if self.state != ChainSyncingState::Paused {
+            // pre-emptively request more blocks from peers whilst we process current blocks,
+            self.send_range_request(network, current_peer);
+
             // Try and process batches sequentially in the ordered list.
             let current_process_id = self.to_be_processed_id;
             for batch in self
@@ -226,20 +251,27 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
                         // the batch was empty, progress to the next block
                         self.to_be_processed_id += 1;
                         continue;
-                    } else if process_batch(chain.clone(), batch) {
-                        info!(log, "Blocks Processed"; "current_slot" => batch.end_slot);
-                        // batch was successfully processed
-                        self.last_processed_id;
-                        self.to_be_processed_id += 1;
                     } else {
-                        // batch processing failed
-                        // this could be because this batch is invalid, or a previous invalidated batch
-                        // is invalid. We need to find out which and downvote the peer that has sent us
-                        // an invalid batch.
+                        warn!(log, "Processing batch"; "batch_id" => batch.id);
+                        match process_batch(chain.clone(), batch, log) {
+                            Ok(_) => {
+                                info!(log, "Blocks Processed"; "current_slot" => batch.end_slot);
+                                // batch was successfully processed
+                                self.last_processed_id = self.to_be_processed_id;
+                                self.to_be_processed_id += 1;
+                            }
+                            Err(e) => {
+                                warn!(log, "Block processing error"; "error"=> format!("{:?}", e));
+                                // batch processing failed
+                                // this could be because this batch is invalid, or a previous invalidated batch
+                                // is invalid. We need to find out which and downvote the peer that has sent us
+                                // an invalid batch.
 
-                        // firstly remove any validated batches
-                        self.handle_invalid_batch(chain, network);
-                        break;
+                                // firstly remove any validated batches
+                                self.handle_invalid_batch(chain, network);
+                                break;
+                            }
+                        }
                     }
                 } else {
                     // there are no more batches to be processed, end
@@ -258,7 +290,6 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
                 // chain is completed
                 true
             } else {
-                self.send_range_request(network, current_peer);
                 // chain is not completed
                 false
             }
@@ -327,11 +358,13 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
 
         // Only important if the local head is more than a batch worth of blocks ahead of
         // what this chain believes is downloaded
-        if let Some(batches_ahead) = local_finalized_slot
+        let batches_ahead = local_finalized_slot
             .as_u64()
             .saturating_sub(self.start_slot.as_u64() + self.last_processed_id * BLOCKS_PER_REQUEST)
             .checked_rem(BLOCKS_PER_REQUEST)
-        {
+            .expect("BLOCKS_PER_REQUEST cannot be 0");
+
+        if batches_ahead != 0 {
             // there are `batches_ahead` whole batches that have been downloaded by another
             // chain. Set the current processed_batch_id to this value.
             debug!(log, "Updating chains processed batches"; "old_completed_slot" => self.start_slot + self.last_processed_id*BLOCKS_PER_REQUEST, "new_completed_slot" => self.start_slot + (self.last_processed_id + batches_ahead)*BLOCKS_PER_REQUEST);
@@ -421,13 +454,14 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
     }
 
     fn get_next_batch(&mut self, peer_id: PeerId) -> Option<Batch<T::EthSpec>> {
-        let batch_start_slot = self.start_slot + self.to_be_downloaded_id * BLOCKS_PER_REQUEST;
+        let batch_start_slot =
+            self.start_slot + self.to_be_downloaded_id.saturating_sub(1) * BLOCKS_PER_REQUEST;
         if batch_start_slot > self.target_head_slot {
             return None;
         }
         let batch_end_slot = std::cmp::min(
-            batch_start_slot + BLOCKS_PER_REQUEST - 1,
-            self.target_head_slot,
+            batch_start_slot + BLOCKS_PER_REQUEST,
+            self.target_head_slot.saturating_add(1u64),
         );
 
         let batch_id = self.to_be_downloaded_id;
@@ -450,9 +484,109 @@ impl<T: BeaconChainTypes> SyncingChain<T> {
     }
 }
 
+// Helper function to process block batches which only consumes the chain and blocks to process
 fn process_batch<T: BeaconChainTypes>(
-    _chain: Weak<BeaconChain<T>>,
-    _batch: &Batch<T::EthSpec>,
-) -> bool {
-    true
+    chain: Weak<BeaconChain<T>>,
+    batch: &Batch<T::EthSpec>,
+    log: &Logger,
+) -> Result<(), String> {
+    for block in &batch.downloaded_blocks {
+        if let Some(chain) = chain.upgrade() {
+            let processing_result = chain.process_block(block.clone());
+
+            if let Ok(outcome) = processing_result {
+                match outcome {
+                    BlockProcessingOutcome::Processed { block_root } => {
+                        // The block was valid and we processed it successfully.
+                        trace!(
+                            log, "Imported block from network";
+                            "slot" => block.slot,
+                            "block_root" => format!("{}", block_root),
+                        );
+                    }
+                    BlockProcessingOutcome::ParentUnknown { parent } => {
+                        // blocks should be sequential and all parents should exist
+                        trace!(
+                            log, "Parent block is unknown";
+                            "parent_root" => format!("{}", parent),
+                            "baby_block_slot" => block.slot,
+                        );
+                        return Err(format!(
+                            "Block at slot {} has an unknown parent.",
+                            block.slot
+                        ));
+                    }
+                    BlockProcessingOutcome::BlockIsAlreadyKnown => {
+                        // this block is already known to us, move to the next
+                        debug!(
+                            log, "Imported a block that is already known";
+                            "block_slot" => block.slot,
+                        );
+                    }
+                    BlockProcessingOutcome::FutureSlot {
+                        present_slot,
+                        block_slot,
+                    } => {
+                        if present_slot + FUTURE_SLOT_TOLERANCE >= block_slot {
+                            // The block is too far in the future, drop it.
+                            trace!(
+                                log, "Block is ahead of our slot clock";
+                                "msg" => "block for future slot rejected, check your time",
+                                "present_slot" => present_slot,
+                                "block_slot" => block_slot,
+                                "FUTURE_SLOT_TOLERANCE" => FUTURE_SLOT_TOLERANCE,
+                            );
+                            return Err(format!(
+                                "Block at slot {} is too far in the future",
+                                block.slot
+                            ));
+                        } else {
+                            // The block is in the future, but not too far.
+                            trace!(
+                                log, "Block is slightly ahead of our slot clock, ignoring.";
+                                "present_slot" => present_slot,
+                                "block_slot" => block_slot,
+                                "FUTURE_SLOT_TOLERANCE" => FUTURE_SLOT_TOLERANCE,
+                            );
+                        }
+                    }
+                    BlockProcessingOutcome::WouldRevertFinalizedSlot { .. } => {
+                        trace!(
+                            log, "Finalized or earlier block processed";
+                            "outcome" => format!("{:?}", outcome),
+                        );
+                        // block reached our finalized slot or was earlier, move to the next block
+                    }
+                    BlockProcessingOutcome::GenesisBlock => {
+                        trace!(
+                            log, "Genesis block was processed";
+                            "outcome" => format!("{:?}", outcome),
+                        );
+                    }
+                    _ => {
+                        warn!(
+                            log, "Invalid block received";
+                            "msg" => "peer sent invalid block",
+                            "outcome" => format!("{:?}", outcome),
+                        );
+                        return Err(format!("Invalid block at slot {}", block.slot));
+                    }
+                }
+            } else {
+                warn!(
+                    log, "BlockProcessingFailure";
+                    "msg" => "unexpected condition in processing block.",
+                    "outcome" => format!("{:?}", processing_result)
+                );
+                return Err(format!(
+                    "Unexpected block processing error: {:?}",
+                    processing_result
+                ));
+            }
+        } else {
+            return Ok(()); // terminate early due to dropped beacon chain
+        }
+    }
+
+    Ok(())
 }
